ref: BookmarkButton and PremiumMark moved to components
