+++ src/components/offer-header/offer-header.tsx
@@ -1,6 +1,6 @@
 import { memo } from 'react';
-import PremiumMark from '../../ui/premium-mark/premium-mark';
 import { capitalize, convertRating } from '../../utils/utils';
+import PremiumMark from '../premium-mark/premium-mark';
 
 // ^======================== OfferHeader ========================^ //
 type OfferHeaderProps = {
